scraper/requirements.txt
beautifulsoup4==4.4.1
decorator==4.0.9
httplib2==0.9.2
lxml==3.6.0
oauth2client==2.0.2
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-http-client==1.2.3
requests==2.9.1
rsa==3.4.2
sendgrid==2.2.1
simplejson==3.8.2
six==1.10.0
smtpapi==0.3.1
sparkpost==1.1.1
uritemplate==0.6
validators==0.10
web.py==0.37
robobrowser==0.5.3
nltk==3.2.1
numpy==1.11.0
python-dateutil==2.5.3
cssselect==0.9.2
